app/requirements.txt
fastapi==0.115.11
numpy==2.2.5
uvicorn==0.30.1
sqlalchemy==2.0.31
sqlmodel==0.0.19
pydantic==2.11.3
alembic==1.12.0
pydantic_settings==2.9.0
#psycopg2
psycopg  #==3.1.10
psycopg[binary,pool]
asyncpg==0.30.0
starlette==0.40.0
py-linq==1.4.0
pillow==11.2.1
bcrypt==4.3.0
pika==1.3.2
python-multipart==0.0.20
huggingface_hub[hf_xet]
passlib==1.7.4
python-jose[cryptography]==3.4.0
jinja2==3.1.2
pytest==8.3.5